looked into and found a new way to do stats with multiple types of modifiers by making a Stat class and turning them all into this Stat class. ran a bunch of tests and got it mostly working. ive deviated from the video a bit as a already have equipment made and dont want to have to remake it from scratch. equipping/unequipping items that give u health also no longer change ur current health. equipment stats are now applied through an event in equipment manager

+++ Assets/Scripts/Items/Accessories.cs
@@ -28,7 +28,7 @@ public class Accessories : Items
 
 	[Header("Accessory Damaging/healing")]
 	public int bonusMana;
-	public int bonusPercentageValue;
+	public float bonusPercentageValue;
 
 	public DamageTypeToBoost damageTypeToBoost;
 	public enum DamageTypeToBoost
